Updated CMake to fix dependency crisis
We are now using a set version of each dependency (except system
libraries like libcurl and libsystemd) and building them with the
library.

There needs to be work done to compile the dependencies
statically to the connections library and have the connections library
only compile static or shared depending on the CMake variable
`BUILD_SHARED_LIBS`

If at some point we add an install target, some headers for dependencies
will need to be installed with this project to the system, one being
Abseil for example. On linux /usr/local/include is not managed by the
main package manager so in most cases it *should* be safe to install
them here. This way there is less likely of a conflict to happen.