Changed Protobuf to generate libraries
Using the `protobuf_generate` function with the TARGET option does nto
compile an actual library, instead adding sources to the build for each
dependency. This causes linking errors when linking to a final binary or
library. This changes it to compile a library (either static or shared)
for each protobuf target.