Fixed `linux` macro defined by Protobuf
The temporary fix completely removed the macro definition. This is not
the best approach as something may need to use that macro in Protobuf,
and resetting it will not allow that to work. This new implementation
uses `push_macro` and `pop_macro` to store and restore the `linux` macro
before and after we remove it (due to the conflicting namespace
declaration)